+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed:
+- Reduce clutter in entry card
+
+## [0.8.269] - 2023-02-20
+### Changed:
 - Replace modal_bottom_sheet lib with Flutter's own implementation
 - Replace badges lib with Flutter's own implementation